requirements.txt
Flask==3.0.0
psycopg2-binary==2.9.9
openpyxl==3.1.2
reportlab==4.0.7
Werkzeug==3.0.1
gunicorn==21.2.0